- The data type is changed from double to long - The @Table annotation is added to the user class - Values are added to the database with liquibase

+++ src/main/java/com/titamedia/challenge/domain/Debt.java
@@ -37,6 +37,6 @@ public class Debt {
     private List<Payment> payments;
     @Temporal(TemporalType.TIMESTAMP)
     private LocalDateTime createdAt;
-    private Double initialAmount;
+    private Long initialAmount;
     private Integer installments;
 }
