+++ src/main/java/com/titamedia/challenge/dto/DebtSummaryDto.java
@@ -3,8 +3,8 @@ package com.titamedia.challenge.dto;
 import com.titamedia.challenge.domain.DebtStatus;
 
 public record DebtSummaryDto(Integer id,
-                             Double paidAmount,
-                             Double pendingAmount,
+                             Long paidAmount,
+                             Long pendingAmount,
                              Integer paidInstallments,
                              Integer pendingInstallments,
                              DebtStatus status) {
